abstracted the styles
